axon7: sepolicy: Actually fix WiFi Mac address
The kernel attempts to read wifimac.dat which
can't be done if it doesn't exist. In the not
so rare case wifimac.dat doesn't exist, the
kernel attempts to create wifimac2.dat. The
current patch assumes wifimac.dat is available
and can be read. On devices without wifimac.dat
it needs to be created and since you can't label
non-existant files users will run into a denial
with persist_file because it's directly writing
to /persist/. So don't attempt to label the mac
file just grant hal_wifi_default proper perms
to store and read the files from /persist/.

Change-Id: I7b15b996db7784156d7bdeabfb9125a69d57a9b5